Display PixelDust and Android versions
Set the title lines to the PixelDust and Android versions and render
them between the logo and menu.

The "Android Recovery" and "Android Fastboot" lines are removed
because the logo already says "Recovery" or "Fastbootd".

Also: Assume the logo and back icon are non-null.
ScreenRecoveryUI::SelectMenu already makes this assumption, and it
avoids questions about how to handle "back" if either the logo or
the back icon is null.

Change-Id: I63bb23e34e52fdb01dd85e793552aa259f46a7de
Signed-off-by: spezi77 <spezi7713@gmx.net>